+++ app.js
@@ -11,7 +11,6 @@ app.use(express.json());
 // define routes
 app.use('/api/contacts', contactsRouter);
 app.use('/api/users', userRouter);
-app.use('/avatars', avatarRouter);
 
 app.use('/', express.static('./public'));
 app.use((_, res) => res.status(404).json({ message: 'Not Found' }));
