+++ middlewares/index.js
@@ -1,4 +1,5 @@
 const validationBody = require('./validationBody');
 const authService = require('./authService');
+const avatarUploadMiddleware = require('./uploadMiddleware');
 
-module.exports = { validationBody, authService };
+module.exports = { validationBody, authService, avatarUploadMiddleware };
