+++ middlewares/uploadMiddleware.js
@@ -1,0 +1,4 @@
+const multer = require('multer');
+const { avatarStorage } = require('../schemas');
+
+module.exports = multer({ storage: avatarStorage });
