+++ public/index.html
@@ -10,11 +10,10 @@
 
   <body>
     <h1>Welcome to Vlad's GoIT Node.js homework project</h1>
-    <a href="/avatars">Avatars</a>
-    <form action="/avatars/upload" enctype="multipart/form-data" method="POST" style="display:block;width:500px;margin:10px;padding:15px;border:1px solid #888888;">
+    <!-- <form action="/avatars/upload" enctype="multipart/form-data" method="POST" style="display:block;width:500px;margin:10px;padding:15px;border:1px solid #888888;">
       <label>Choose avatar:<input type="file" name="avatar" /></label><br/>
       <button type="submit">Upload</button>
-    </form>
+    </form> -->
   </body>
 
 </html>
