+++ routes/index.js
@@ -1,5 +1,4 @@
 const contactsRouter = require('./contactsRouter');
 const userRouter = require('./userRouter');
-const avatarRouter = require('./avatarRouter');
 
-module.exports = { contactsRouter, userRouter, avatarRouter };
+module.exports = { contactsRouter, userRouter };
