+++ schemas/index.js
@@ -1,4 +1,5 @@
 const { contactDbSchema, contactJoiSchemas } = require('./contactSchemas');
 const { userDbSchema, userJoiSchemas } = require('./userSchemas');
+const { avatarStorage } = require('./avatarStorage');
 
-module.exports = { userDbSchema, userJoiSchemas, contactDbSchema, contactJoiSchemas };
+module.exports = { userDbSchema, userJoiSchemas, contactDbSchema, contactJoiSchemas, avatarStorage };
